trading/hyperliquid-bot/requirements.txt
hyperliquid-python-sdk==0.22.0
numpy>=1.24.0
eth-account>=0.10.0
requests>=2.31.0
web3>=7.0.0
